Move AST methods to ast_method
